Renamed and moved PlatformFile.open api to FileKit.nonWeb.openFile api, removed samples about this new api method
