+++ filekit-core/src/iosMain/kotlin/io/github/vinceglb/filekit/FileKit.ios.kt
@@ -15,13 +15,7 @@ import platform.CoreGraphics.CGSizeMake
 import platform.Foundation.NSBundle
 import platform.Foundation.NSData
 import platform.Foundation.create
-import platform.UIKit.UIGraphicsBeginImageContextWithOptions
-import platform.UIKit.UIGraphicsEndImageContext
-import platform.UIKit.UIGraphicsGetImageFromCurrentImageContext
-import platform.UIKit.UIImage
-import platform.UIKit.UIImageJPEGRepresentation
-import platform.UIKit.UIImagePNGRepresentation
-import platform.UIKit.UIImageWriteToSavedPhotosAlbum
+import platform.UIKit.*
 
 public actual val FileKit.projectDir: PlatformFile
     get() = PlatformFile(nsUrl = NSBundle.mainBundle.bundleURL)
@@ -77,4 +71,15 @@ private fun UIImage.scaleToSize(newWidth: Int, newHeight: Int): UIImage? {
     val resizedImage = UIGraphicsGetImageFromCurrentImageContext()
     UIGraphicsEndImageContext()
     return resizedImage
+}
+
+public actual fun FileKit.openFile(
+    file: PlatformFile
+) {
+    val documentController = UIDocumentInteractionController()
+    documentController.URL = file.nsUrl
+    documentController.delegate()
+    documentController.presentPreviewAnimated(
+        animated = true
+    )
 }
