--- filekit-core/src/iosMain/kotlin/io/github/vinceglb/filekit/PlatformFile.ios.kt
@@ -1,13 +1,0 @@
-package io.github.vinceglb.filekit
-
-import platform.UIKit.UIDocumentInteractionController
-
-public actual fun PlatformFile.open() {
-    val documentController = UIDocumentInteractionController()
-    documentController.URL = this.nsUrl
-    documentController.delegate()
-    documentController.presentPreviewAnimated(
-        animated = true
-    )
-}
-
