+++ filekit-core/src/nonWebMain/kotlin/io/github/vinceglb/filekit/FileKit.nonWeb.kt
@@ -44,3 +44,7 @@ public suspend fun FileKit.compressImage(
     maxHeight = maxHeight,
     imageFormat = imageFormat
 )
+
+public expect fun FileKit.openFile(
+    file: PlatformFile
+)
