+++ filekit-core/src/webMain/kotlin/io/github/vinceglb/filekit/PlatformFile.web.kt
@@ -6,7 +6,3 @@ public actual suspend fun PlatformFile.readString(): String =
 public actual fun PlatformFile.startAccessingSecurityScopedResource(): Boolean = true
 
 public actual fun PlatformFile.stopAccessingSecurityScopedResource() {}
-
-public actual fun PlatformFile.open() {
-    TODO("Not implemented yet, to do it")
-}
